T1ch2 little fix

+++ src/main/java/com/safedjio/arraytask/specification/Impl/IdSpecification.java
@@ -4,9 +4,9 @@ import com.safedjio.arraytask.entity.CustomArray;
 import com.safedjio.arraytask.specification.CustomSpecification;
 
 public class IdSpecification implements CustomSpecification {
-    private final int wantedId;
+    private final Long wantedId;
 
-    public IdSpecification(int wantedId){
+    public IdSpecification(Long wantedId){
         this.wantedId = wantedId;
     }
